cleanup and refactoring
