+++ src/main/java/os/failsafe/executor/task/FailedTask.java
@@ -35,6 +35,7 @@ public class FailedTask {
     private final TaskId id;
     private final String parameter;
     private final String name;
+    //TODO unused
     private final LocalDateTime startTime;
     private final Long version;
     private final LocalDateTime failTime;
